@@ -15,21 +15,12 @@
         <project.reporting.outputEncoding>${encoding.type}</project.reporting.outputEncoding>
         <file.encoding>${encoding.type}</file.encoding>
         <!-- JDK version project -->
-        <java.version>11</java.version>
+        <java.version>21</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
-        <!-- Dependencies version -->
-        <junit-jupiter.version>5.7.1</junit-jupiter.version>
-        <selenium-java.version>4.0.0</selenium-java.version>
-        <maven-jar-plugin.version>3.3.0</maven-jar-plugin.version>
-        <logback-classic.version>1.2.6</logback-classic.version>
-        <junit-jupiter-api>5.9.2</junit-jupiter-api>
-        <selenium-http-jdk-client>4.5.0</selenium-http-jdk-client>
-        <logback-classic>1.2.6</logback-classic>
-        <slf4j-api>1.7.30</slf4j-api>
         <allure.version>2.26.0</allure.version>
         <aspectj.version>1.9.22</aspectj.version>
-        <maven-surefire.version>3.0.0-M5</maven-surefire.version>
+        <!-- Dependencies version -->
     </properties>
 
     <dependencyManagement>
