@@ -3,12 +3,12 @@ package ru.framework.pages.taskone;
 import io.qameta.allure.Step;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.framework.BasePage;
+import org.junit.jupiter.api.Assertions;
 import ru.framework.pages.taskthree.LaptopsPage;
 import ru.framework.utils.PropsConst;
 
